requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
future==0.18.2
idna==2.10
oauthlib==3.1.1
python-twitter==3.5
requests==2.25.1
requests-oauthlib==1.3.0
soupsieve==2.2.1
urllib3==1.26.6
